Separate Qt-specific pieces into their own directory.
