+++ glue_aladin/qt/catalog_layer_widget.py
