+++ glue_aladin/qt/tests/test_viewer.py
